pyproject.toml
[tool.poetry]
name = "pythonista-bot"
version = "0.1.0"
description = "The bot that runs the Pythonista guild!"
authors = ["Alex Nørgaard <Umbra@AbstractUmbra.dev>"]
license = "MPL2.0"
readme = "README.md"
# packages = [{include = "pythonista_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "*"
aiohttp = "*"
asyncpg = "*"
toml = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
"asyncpg-stubs" = "*"

[tool.black]
line-length = 125

[tool.isort]
profile = "black"
lines_after_imports = 2
line_length = 125
combine_as_imports = true
combine_star = true


[tool.pyright]
useLibraryCodeForTypes = true
typeCheckingMode = "strict"
pythonVersion = "3.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
